Use different test class for installed service tests
